requirements.txt
flask
numpy==1.15.0
pillow
tensorflow==1.12.0

